support-case-mcp/pyproject.toml
[project]
name = "support-case-mcp"
version = "0.1.0"
description = "MCP Server for Silicon Labs Support Use Cases (Salesforce)"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "simple-salesforce>=1.12.5",
    "python-dotenv>=1.0.0"
]

[tool.setuptools]
packages = ["support_case_mcp"]
